Move MainActivity to 'view' package
